Embedded the items that make up the search form in a new parent view
The idea here is not to put these elements as direct descendents of the top-level view. Instead I'm making different screens, or pages, of views. This allows me to easily switch between different layouts, like, for example, a welcome screen when you first boot up, or another one if you have no drivers.